backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
echo openai==1.30.3>>requirements.txt
echo python-dotenv==1.0.1>>requirements.txt
openai==1.30.3
httpx==0.27.0
python-dotenv==1.0.1